# Conflicts:
#	settings.gradle
#	tkpdpayment/build.gradle
#	tkpdpayment/proguard-rules.pro
#	tkpdpayment/src/androidTest/java/com/tokopedia/payment/ExampleInstrumentedTest.java
#	tkpdpayment/src/main/AndroidManifest.xml
#	tkpdpayment/src/main/res/values/strings.xml
#	tkpdpayment/src/test/java/com/tokopedia/payment/ExampleUnitTest.java

+++ tkpddiscovery/src/main/java/com/tokopedia/discovery/fragment/ProductFragment.java
@@ -12,7 +12,6 @@ import android.support.design.widget.Snackbar;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.Display;
 import android.view.View;
 import android.view.WindowManager;
@@ -37,8 +36,8 @@ import com.tokopedia.core.var.RecyclerViewItem;
 import com.tokopedia.core.var.TkpdState;
 import com.tokopedia.discovery.activity.BrowseProductActivity;
 import com.tokopedia.discovery.adapter.DefaultCategoryAdapter;
-import com.tokopedia.discovery.adapter.RevampCategoryAdapter;
 import com.tokopedia.discovery.adapter.ProductAdapter;
+import com.tokopedia.discovery.adapter.RevampCategoryAdapter;
 import com.tokopedia.discovery.interfaces.FetchNetwork;
 import com.tokopedia.discovery.presenter.FragmentDiscoveryPresenter;
 import com.tokopedia.discovery.presenter.FragmentDiscoveryPresenterImpl;
@@ -170,18 +169,18 @@ public class ProductFragment extends BaseFragment<FragmentDiscoveryPresenter>
         }else{
             productAdapter.setIsLoading(false);
         }
-        if (getActivity() !=null && getActivity() instanceof BrowseProductActivity) {
+        if (getActivity() != null && getActivity() instanceof BrowseProductActivity) {
             BrowseProductActivityModel browseModel = ((BrowseProductActivity) getActivity()).getBrowseProductActivityModel();
-            if(browseModel.getHotListBannerModel()!=null){
+            if (browseModel.getHotListBannerModel() != null) {
                 HotListBannerModel bannerModel = browseModel.getHotListBannerModel();
-                if(bannerModel.query.shop_id.isEmpty()){
+                if (bannerModel.query.shop_id.isEmpty()) {
                     presenter.getTopAds(productAdapter.getTopAddsCounter(), TAG, getActivity(), spanCount);
                 }
-            } else if (model.size() > 0){
+            } else if (model.size() > 0) {
                 presenter.getTopAds(productAdapter.getTopAddsCounter(), TAG, getActivity(), spanCount);
             }
 
-            if (totalProduct>0)
+            if (totalProduct > 0)
                 browseModel.setTotalDataCategory(NumberFormat.getNumberInstance(Locale.US)
                         .format(totalProduct.longValue()).replace(',', '.'));
             productAdapter.incrementPage();
@@ -214,11 +213,11 @@ public class ProductFragment extends BaseFragment<FragmentDiscoveryPresenter>
             productAdapter.setIsLoading(false);
         }
 
-        if (getActivity()!=null && getActivity() instanceof  BrowseProductActivity) {
+        if (getActivity() != null && getActivity() instanceof BrowseProductActivity) {
             BrowseProductActivityModel browseModel = ((BrowseProductActivity) getActivity()).getBrowseProductActivityModel();
-            if(browseModel.getHotListBannerModel()!=null){
+            if (browseModel.getHotListBannerModel() != null) {
                 HotListBannerModel bannerModel = browseModel.getHotListBannerModel();
-                if(bannerModel.query.shop_id.isEmpty()){
+                if (bannerModel.query.shop_id.isEmpty()) {
                     presenter.getTopAds(getPage(ProductFragment.TAG), TAG, getActivity(), spanCount);
                 }
             } else {
@@ -435,15 +434,15 @@ public class ProductFragment extends BaseFragment<FragmentDiscoveryPresenter>
     @Override
     public void addCategoryHeader(Data categoryHeader) {
         isHasCategoryHeader = true;
-        if (getActivity()!=null && getActivity() instanceof  BrowseProductActivity) {
+        if (getActivity() != null && getActivity() instanceof BrowseProductActivity) {
             BrowseProductActivityModel browseModel = ((BrowseProductActivity) getActivity()).getBrowseProductActivityModel();
-            if (categoryHeader.getRevamp() !=null && categoryHeader.getRevamp()) {
+            if (categoryHeader.getRevamp() != null && categoryHeader.getRevamp()) {
                 productAdapter.addCategoryRevampHeader(
-                        new ProductAdapter.CategoryHeaderRevampModel(categoryHeader,getActivity(),getCategoryWidth(),
-                                this,browseModel.getTotalDataCategory(), this));
+                        new ProductAdapter.CategoryHeaderRevampModel(categoryHeader, getActivity(), getCategoryWidth(),
+                                this, browseModel.getTotalDataCategory(), this));
             } else {
                 productAdapter.addCategoryHeader(
-                        new ProductAdapter.CategoryHeaderModel(categoryHeader,getActivity(),getCategoryWidth(),
+                        new ProductAdapter.CategoryHeaderModel(categoryHeader, getActivity(), getCategoryWidth(),
                                 this, browseModel.getTotalDataCategory(), this));
             }
             productAdapter.notifyDataSetChanged();
@@ -470,7 +469,7 @@ public class ProductFragment extends BaseFragment<FragmentDiscoveryPresenter>
         Point size = new Point();
         display.getSize(size);
         int width = size.x;
-        return (int) (width / 2);
+        return width / 2;
     }
 
     @Override
